feat: vamm events id

+++ src/interfaces/IPoolModule.sol
@@ -46,7 +46,7 @@ interface IPoolModule is IERC165 {
     function initiateDatedMakerOrder(
         uint128 accountId,
         uint128 marketId,
-        uint256 maturityTimestamp,
+        uint32 maturityTimestamp,
         int24 tickLower,
         int24 tickUpper,
         int128 liquidityDelta
